+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+<a name="2.6.0"></a>
+# [2.6.0] (2017-00-00)
+
+### Features
+- none
+
+### UI/UX Changes
+- Moving go to current button inside the navigation component (#commit).
+
+### Bug Fixed
+- none
+
+### Breaking Changes
+- Go to current button moved from input element to the navigation component (#commit).
+
+
+
 <a name="2.5.1"></a>
 # [2.5.1] (2017-10-12)